python-observability-app/backend/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.9
Flask-CORS==4.0.0
newrelic
